Refactor src to use relative paths
* Uses React apps built in dotenv integration to allow loading of modules locally (without having to use relative path with ‘import`).

* For this to work it needs a `.env` file at `ui/.env` with the following contents:

    NODE_PATH=src/

We would need to put this in ‘aleph.env’, or otherwise inject it into the Docker image in a similar way.

Those working in an IDE who want auto-completion would also want to have it in the .env too (and might involve some configuration - see http://cuneyt.aliustaoglu.biz/en/absolute-imports-and-intellisense-with-create-react-app/).

There are other approaches to ‘not using a symlink’ but this is the one create-react-app supports out of the box and that seems to work with both development and production builds (on client and server) without breaking things like running scripts for react-intl or tests.

Changing NODE_PATH is generally discouraged by the node team - as is using absolute paths to require modules - but the options to do this behaviour are more limited with create-react-app.

Many of the alternative ways are more fragile, platform specific or, in our specific case, incompatible with our project without additional refactoring. Suggest we (the Alpeh core team) chat about them.